IO, documentation

+++ src/main/java/Interfaces/MouseEventHandler.java
@@ -3,5 +3,5 @@ package Interfaces;
 public interface MouseEventHandler {
 
     void processMouseEvent(int key, int action);
-    void processMouseMovement(int X, int Y);
+    void processMouseMovement(double X, double Y);
 }
